build : upgrade bonus parsing fix : init problem norm : begin to norm bonus

+++ srcs_bonus/color_bonus.c
@@ -8,12 +8,12 @@ void	limit_color(t_argb *color)
 	color->b = fmin(255, fmax(0, color->b));
 }
 
-unsigned int encode_argb(t_argb color)
+uint8_t	encode_argb(t_argb color)
 {
-	const uint8_t alpha = color.a;
-   	const uint8_t red = color.r;
-   	const uint8_t green = color.g;
-	const uint8_t blue = color.b;
+	const uint8_t	alpha = color.a;
+	const uint8_t	red = color.r;
+	const uint8_t	green = color.g;
+	const uint8_t	blue = color.b;
 
 	return (alpha << 24 | red << 16 | green << 8 | blue);
 }
@@ -21,7 +21,7 @@ unsigned int encode_argb(t_argb color)
 t_argb	ease_color(t_argb color, uint8_t ease)
 {
 	t_argb	result;
-	
+
 	result.a = ease - color.a;
 	result.r = ease - color.r;
 	result.g = ease - color.g;
@@ -56,7 +56,7 @@ t_argb	add_colors(t_argb color1, t_argb color2)
 
 t_argb	apply_brightness(t_argb color)
 {
-	const float brightness = (float)color.a / 255;
+	const float	brightness = (float)color.a / 255;
 	t_argb		real_color;
 
 	real_color.a = 255;
@@ -68,7 +68,7 @@ t_argb	apply_brightness(t_argb color)
 
 t_argb	invert_color(t_argb color)
 {
-	t_argb new_color;
+	t_argb	new_color;
 
 	new_color.a = 255 - color.a;
 	new_color.r = 255 - color.r;
